requirements.txt
python-telegram-bot[job-queue]>=20.0
pandas>=1.3.0
cryptography>=36.0.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
